project/backend/requirements.txt
asgiref==3.9.1
boto3==1.35.86
botocore==1.35.86
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
django-cors-headers==4.9.0
django-storages==1.14.4
djangorestframework==3.16.1
docopt==0.6.2
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.3.0
pipreqs==0.4.13
psycopg2-binary==2.9.10
python-dotenv==1.1.1
reportlab==4.4.3
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
yarg==0.1.10
